drivers: marvell: gsern: Fix declaration of few tuning parmeters
Some of tuning parameters are declared as uint8_t but used as int.
Fix the incompatibility. Due to this, serdes is not tuned correctly.

Reviewed-on: https://sj1git1.cavium.com/13133
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>

+++ include/plat/marvell/octeontx/otx2/plat_board_cfg.h
@@ -78,10 +78,10 @@ typedef struct cgx_config {
 } cgx_config_t;
 
 typedef struct qlm_config {
-	uint8_t lane_rx_polarity[MAX_LANES_PER_QLM];
-	uint8_t lane_tx_polarity[MAX_LANES_PER_QLM];
-	uint8_t lane_rx_prevga_gn_adapt[MAX_LANES_PER_QLM];
-	uint8_t lane_rx_prevga_gn_ovrd[MAX_LANES_PER_QLM];
+	int lane_rx_polarity[MAX_LANES_PER_QLM];
+	int lane_tx_polarity[MAX_LANES_PER_QLM];
+	int lane_rx_prevga_gn_adapt[MAX_LANES_PER_QLM];
+	int lane_rx_prevga_gn_ovrd[MAX_LANES_PER_QLM];
 } qlm_config_t;
 
 typedef struct mcc_ras_config {
